DynamicVolumeLines: Fix warnings

+++ modules/DynamicVolumeLines/iANDimImagePointer.cpp
@@ -29,14 +29,14 @@ iANDimImagePointer::iANDimImagePointer(unsigned int dimensionCount, unsigned int
 {
 	// sum up all elements in their hypercube and return the value divided by the count of all elements
 
-	for (int i = 0; i < _values.size(); i++) {
+	for (size_t i = 0; i < _values.size(); i++)
+	{
 		std::vector<double> record = _values[i];
-
 		itk::IndexValueType * index = new itk::IndexValueType[dimensionCount];
-		for (int j = 0; j < record.size(); j++) {
-
+		for (size_t j = 0; j < record.size(); j++)
+		{
 			double value = record[j];
-			index[j] = (int)value;
+			index[j] = static_cast<int>(value);
 		}
 
 		Key indexKey;
@@ -55,8 +55,10 @@ iANDimImagePointer::~iANDimImagePointer()
 unsigned int iANDimImagePointer::getMaxDim()
 {
 	unsigned int maxDim = 0;
-	for (int i = 0; i < _dimensionCount; i++) {
-		if (_dimSize[i] > maxDim) {
+	for (unsigned int i = 0; i < _dimensionCount; i++)
+	{
+		if (_dimSize[i] > maxDim)
+		{
 			maxDim = _dimSize[i];
 		}
 	}
@@ -74,7 +76,8 @@ double iANDimImagePointer::getDensityAt(const itk::IndexValueType * index, unsig
 	indexKey.indizes = index;
 
 	auto density = _densityMap.find(indexKey);
-	if (density != _densityMap.end()) {
+	if (density != _densityMap.end())
+	{
 		return density->second;
 	}
 	return 0.0;
@@ -108,7 +111,8 @@ double iANDimImagePointer::getDensityAt(const itk::IndexValueType * index, unsig
 unsigned int iANDimImagePointer::getLinearizedDimSize()
 {
 	unsigned int ret = 1;
-	for (int i = 0; i < _dimensionCount; i++) {
+	for (unsigned int i = 0; i < _dimensionCount; i++)
+	{
 		ret *= _dimSize[i];
 	}
 	return ret;
